requirements.txt
streamlit
pandas
pymupdf
numpy
opencv-python-headless
pillow
paddleocr
paddlepaddle
openai
xlsxwriter